DE112: To solve the issue with the back link not working consistantly on the standards page we have persisted the link through the return URL from the add/remove to shortlist calls.
Now the standards page sends its link back url to the shortlist controller which (if possible) will return the same URL via a URL parameter to the standards page so that the back link is maintained after the standard has been add/removed  rom the shortlist via the link on the standards page.

If the link is not provided or the previous page link cannot be found in the request then the default serach page will be returned instead.